+++ src/app/cms/product/page.jsx
@@ -4,9 +4,12 @@ import { useSelector } from "react-redux";
 import AuthMainLayout from "../../layouts/auth/AuthMainLayout";
 import api from "@/utils/api";
 import { toast } from "react-toastify";
+import { getCmsAccess, getDeletePermissionMessage } from "@/utils/cmsAccess";
 
 const CmsDesignGallery = () => {
+    const user = useSelector((state) => state.auth.user);
     const authToken = useSelector((state) => state.auth.authToken);
+    const { canDelete } = getCmsAccess(user);
     const [pagesList, setPagesList] = useState();
     const [loading, setLoading] = useState(false);
     const [formData, setFormData] = useState({
@@ -182,6 +185,11 @@ const CmsDesignGallery = () => {
     }
 
     const deleteHandler = async (id) => {
+        if (!canDelete) {
+            toast.error(getDeletePermissionMessage("this product"));
+            return;
+        }
+
         if (window.confirm("Are you sure you want to delete this product?")) {
             try {
                 const response = await api.delete(`/cms-parent-child/${id}`, {
@@ -252,7 +260,7 @@ const CmsDesignGallery = () => {
                                             <button onClick={() => handleEditClick(item)} type="button" className="read_morebtn" data-bs-toggle="modal" data-bs-target="#editNewpageModal">
                                                 Edit
                                             </button>
-                                            <button className="ms-2 btn btn-danger" onClick={() => deleteHandler(item.id)}>Delete</button>
+                                            {canDelete && <button className="ms-2 btn btn-danger" onClick={() => deleteHandler(item.id)}>Delete</button>}
                                         </td>
                                     </tr>
                                 ))}
